refactor(issue-181): VantageLPPage → Portfolio統合 / /vantage-lp → /portfolioリダイレクト
